fabric_service/fabric_client/requirements.txt
grpcio
protobuf == 3.19.1
pycryptodomex
toml
fabric-sdk-py